refactor: rename mount/ package to disks/
Rename the mount package to disks to better describe the domain
(physical disk management: device detection, LUKS attach/detach,
mount/umount, auth rules). This avoids confusion with the
config-level "volumes" concept (LocalVolume/RemoteVolume).

Mechanical rename of directory + all imports across the codebase.
