+++ nbkp/disks/strategy.py
