+++ tests/_docker_fixtures/__init__.py
@@ -21,7 +21,7 @@ from nbkp.config import (
 from nbkp.config.epresolution import ResolvedEndpoints
 from nbkp.remote.resolution import resolve_all_endpoints
 from nbkp.config import LuksEncryptionConfig, MountConfig
-from nbkp.mount.strategy import DirectMountStrategy
+from nbkp.disks.strategy import DirectMountStrategy
 from nbkp.remote.testkit.docker import (  # noqa: F401
     DOCKER_DIR,
     LUKS_MAPPER_NAME,
